feat: damage number 적용, Player, Monster에 IDamageable 적용

+++ Assets/Scripts/Skill/IStatsProvider.cs
@@ -41,6 +41,6 @@ namespace SlayerLegend.Skill
     // 데미지를 받을 수 있는 대상 인터페이스
     public interface IDamageable
     {
-        void TakeDamage(float damage);
+        void TakeDamage(double damage);
     }
 }
